Move reducer to its own package
